# openapi_chatgpt_v2.yaml
openapi: 3.1.0
info:
  title: Life Fractal Intelligence API
  description: |
    🌀 Transform your goals into beautiful fractal art!
    
    Features:
    - 🐾 Virtual Pet Companions (8 species)
    - 🎨 12+ Fractal Types with Sacred Geometry
    - 📊 Goal Visualization for Aphantasia
    - 🧠 Neurodivergent-Friendly Design
    
    Built for minds that work differently.
  version: 2.0.0
servers:
  - url: https://planner-1-pyd9.onrender.com
    description: Production server
paths:
...
  /chatgpt/fractals/generate:
    post:
      operationId: generateFractal
      summary: Generate a fractal image
      description: Create a beautiful fractal image. Returns base64 PNG.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  description: Fractal type to generate
                  enum:
                    - mandelbrot
                    - julia
                    - golden_spiral
                    - flower_of_life
                    - burning_ship
                    - phoenix
                    - sierpinski
                    - dragon
                    - koch_snowflake
                    - fibonacci
                    - lorenz
                    - metatron
                size:
                  type: integer
                  description: Image size (256-2048)
                  minimum: 256
                  maximum: 2048
                  default: 800
                params:
                  type: object
                  description: Optional fractal parameters
                  properties:
                    zoom:
                      type: number
                    iterations:
                      type: integer
                    c_real:
                      type: number
                    c_imag:
                      type: number
      responses:
        "200":
          description: Fractal generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  fractal_type:
                    type: string
                  fractal_name:
                    type: string
                  meaning:
                    type: string
                  image_data:
                    type: string
                    description: Base64 PNG with data URI prefix
                  message:
                    type: string
...
  /chatgpt/pet/create:
    post:
      operationId: createPet
      summary: Create a virtual pet
      description: Create your fractal companion!
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - species
              properties:
                species:
                  type: string
                  description: Pet species
                  enum:
                    - cat
                    - dog
                    - dragon
                    - phoenix
                    - owl
                    - fox
                    - unicorn
                    - butterfly
                name:
                  type: string
                  description: Pet name
                  maxLength: 50
                  default: Fractal Friend
      responses:
        "200":
          description: Pet created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pet:
                    type: object
                    properties:
                      species:
                        type: string
                      name:
                        type: string
                      level:
                        type: integer
                      happiness:
                        type: integer
                      hunger:
                        type: integer
                      energy:
                        type: integer
                      mood:
                        type: string
                  greeting:
                    type: string
                  message:
                    type: string
...
  /chatgpt/goals/create:
    post:
      operationId: createGoal
      summary: Create a goal
      description: Create a goal and get its fractal type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  maxLength: 200
                type:
                  type: string
                  enum: [karma, dharma]
                  default: karma
                points:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 50
      responses:
        "200":
          description: Goal created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  goal:
                    type: object
                  suggested_fractal:
                    type: object
                  message:
                    type: string
...